+++ src/sdlgfx/constantsmod.c
@@ -24,7 +24,7 @@
 #include <SDL_rotozoom.h>
 
 /* macros used to create each constant */
-#define DEC_CONSTN(x)  PyModule_AddIntConstant(module, #x, (int) x)
+#define DEC_CONSTN(x)  PyModule_AddIntConstant(module, #x, (long) x)
 
 #ifdef IS_PYTHON_3
 PyMODINIT_FUNC PyInit_constants (void)
